+++ sdk/src/main/java/jp/co/freee/accounting/models/ManualJournalUpdateParamsDetails.java
@@ -38,7 +38,7 @@ public class ManualJournalUpdateParamsDetails {
 
   public static final String SERIALIZED_NAME_AMOUNT = "amount";
   @SerializedName(SERIALIZED_NAME_AMOUNT)
-  private Integer amount;
+  private Long amount;
 
   public static final String SERIALIZED_NAME_DESCRIPTION = "description";
   @SerializedName(SERIALIZED_NAME_DESCRIPTION)
@@ -97,7 +97,7 @@ public class ManualJournalUpdateParamsDetails {
 
   public static final String SERIALIZED_NAME_ID = "id";
   @SerializedName(SERIALIZED_NAME_ID)
-  private Integer id;
+  private Long id;
 
   public static final String SERIALIZED_NAME_ITEM_ID = "item_id";
   @SerializedName(SERIALIZED_NAME_ITEM_ID)
@@ -117,15 +117,15 @@ public class ManualJournalUpdateParamsDetails {
 
   public static final String SERIALIZED_NAME_SEGMENT1_TAG_ID = "segment_1_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT1_TAG_ID)
-  private Integer segment1TagId;
+  private Long segment1TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT2_TAG_ID = "segment_2_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT2_TAG_ID)
-  private Integer segment2TagId;
+  private Long segment2TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT3_TAG_ID = "segment_3_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT3_TAG_ID)
-  private Integer segment3TagId;
+  private Long segment3TagId;
 
   public static final String SERIALIZED_NAME_TAG_IDS = "tag_ids";
   @SerializedName(SERIALIZED_NAME_TAG_IDS)
@@ -164,7 +164,7 @@ public class ManualJournalUpdateParamsDetails {
   }
 
 
-  public ManualJournalUpdateParamsDetails amount(Integer amount) {
+  public ManualJournalUpdateParamsDetails amount(Long amount) {
     
     this.amount = amount;
     return this;
@@ -178,12 +178,12 @@ public class ManualJournalUpdateParamsDetails {
   **/
   @ApiModelProperty(example = "10800", required = true, value = "取引金額（税込で指定してください）")
 
-  public Integer getAmount() {
+  public Long getAmount() {
     return amount;
   }
 
 
-  public void setAmount(Integer amount) {
+  public void setAmount(Long amount) {
     this.amount = amount;
   }
 
@@ -233,7 +233,7 @@ public class ManualJournalUpdateParamsDetails {
   }
 
 
-  public ManualJournalUpdateParamsDetails id(Integer id) {
+  public ManualJournalUpdateParamsDetails id(Long id) {
     
     this.id = id;
     return this;
@@ -248,12 +248,12 @@ public class ManualJournalUpdateParamsDetails {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "貸借行ID: 既存貸借行を更新または削除する場合に指定します。IDを指定しない貸借行は、新規行として扱われ追加されます。")
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
@@ -356,7 +356,7 @@ public class ManualJournalUpdateParamsDetails {
   }
 
 
-  public ManualJournalUpdateParamsDetails segment1TagId(Integer segment1TagId) {
+  public ManualJournalUpdateParamsDetails segment1TagId(Long segment1TagId) {
     
     this.segment1TagId = segment1TagId;
     return this;
@@ -371,17 +371,17 @@ public class ManualJournalUpdateParamsDetails {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント１ID")
 
-  public Integer getSegment1TagId() {
+  public Long getSegment1TagId() {
     return segment1TagId;
   }
 
 
-  public void setSegment1TagId(Integer segment1TagId) {
+  public void setSegment1TagId(Long segment1TagId) {
     this.segment1TagId = segment1TagId;
   }
 
 
-  public ManualJournalUpdateParamsDetails segment2TagId(Integer segment2TagId) {
+  public ManualJournalUpdateParamsDetails segment2TagId(Long segment2TagId) {
     
     this.segment2TagId = segment2TagId;
     return this;
@@ -396,17 +396,17 @@ public class ManualJournalUpdateParamsDetails {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント２ID")
 
-  public Integer getSegment2TagId() {
+  public Long getSegment2TagId() {
     return segment2TagId;
   }
 
 
-  public void setSegment2TagId(Integer segment2TagId) {
+  public void setSegment2TagId(Long segment2TagId) {
     this.segment2TagId = segment2TagId;
   }
 
 
-  public ManualJournalUpdateParamsDetails segment3TagId(Integer segment3TagId) {
+  public ManualJournalUpdateParamsDetails segment3TagId(Long segment3TagId) {
     
     this.segment3TagId = segment3TagId;
     return this;
@@ -421,12 +421,12 @@ public class ManualJournalUpdateParamsDetails {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント３ID")
 
-  public Integer getSegment3TagId() {
+  public Long getSegment3TagId() {
     return segment3TagId;
   }
 
 
-  public void setSegment3TagId(Integer segment3TagId) {
+  public void setSegment3TagId(Long segment3TagId) {
     this.segment3TagId = segment3TagId;
   }
