+++ sdk/src/main/java/jp/co/freee/accounting/models/QuotationQuotationContents.java
@@ -82,7 +82,7 @@ public class QuotationQuotationContents {
 
   public static final String SERIALIZED_NAME_SEGMENT1_TAG_ID = "segment_1_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT1_TAG_ID)
-  private Integer segment1TagId;
+  private Long segment1TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT1_TAG_NAME = "segment_1_tag_name";
   @SerializedName(SERIALIZED_NAME_SEGMENT1_TAG_NAME)
@@ -90,7 +90,7 @@ public class QuotationQuotationContents {
 
   public static final String SERIALIZED_NAME_SEGMENT2_TAG_ID = "segment_2_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT2_TAG_ID)
-  private Integer segment2TagId;
+  private Long segment2TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT2_TAG_NAME = "segment_2_tag_name";
   @SerializedName(SERIALIZED_NAME_SEGMENT2_TAG_NAME)
@@ -98,7 +98,7 @@ public class QuotationQuotationContents {
 
   public static final String SERIALIZED_NAME_SEGMENT3_TAG_ID = "segment_3_tag_id";
   @SerializedName(SERIALIZED_NAME_SEGMENT3_TAG_ID)
-  private Integer segment3TagId;
+  private Long segment3TagId;
 
   public static final String SERIALIZED_NAME_SEGMENT3_TAG_NAME = "segment_3_tag_name";
   @SerializedName(SERIALIZED_NAME_SEGMENT3_TAG_NAME)
@@ -453,7 +453,7 @@ public class QuotationQuotationContents {
   }
 
 
-  public QuotationQuotationContents segment1TagId(Integer segment1TagId) {
+  public QuotationQuotationContents segment1TagId(Long segment1TagId) {
     
     this.segment1TagId = segment1TagId;
     return this;
@@ -466,12 +466,12 @@ public class QuotationQuotationContents {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント１ID")
 
-  public Integer getSegment1TagId() {
+  public Long getSegment1TagId() {
     return segment1TagId;
   }
 
 
-  public void setSegment1TagId(Integer segment1TagId) {
+  public void setSegment1TagId(Long segment1TagId) {
     this.segment1TagId = segment1TagId;
   }
 
@@ -499,7 +499,7 @@ public class QuotationQuotationContents {
   }
 
 
-  public QuotationQuotationContents segment2TagId(Integer segment2TagId) {
+  public QuotationQuotationContents segment2TagId(Long segment2TagId) {
     
     this.segment2TagId = segment2TagId;
     return this;
@@ -512,12 +512,12 @@ public class QuotationQuotationContents {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント２ID")
 
-  public Integer getSegment2TagId() {
+  public Long getSegment2TagId() {
     return segment2TagId;
   }
 
 
-  public void setSegment2TagId(Integer segment2TagId) {
+  public void setSegment2TagId(Long segment2TagId) {
     this.segment2TagId = segment2TagId;
   }
 
@@ -545,7 +545,7 @@ public class QuotationQuotationContents {
   }
 
 
-  public QuotationQuotationContents segment3TagId(Integer segment3TagId) {
+  public QuotationQuotationContents segment3TagId(Long segment3TagId) {
     
     this.segment3TagId = segment3TagId;
     return this;
@@ -558,12 +558,12 @@ public class QuotationQuotationContents {
   @javax.annotation.Nullable
   @ApiModelProperty(example = "1", value = "セグメント３ID")
 
-  public Integer getSegment3TagId() {
+  public Long getSegment3TagId() {
     return segment3TagId;
   }
 
 
-  public void setSegment3TagId(Integer segment3TagId) {
+  public void setSegment3TagId(Long segment3TagId) {
     this.segment3TagId = segment3TagId;
   }
